requirements.txt
openai==1.58.1
python-wordpress-xmlrpc==2.3
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.3
beautifulsoup4==4.12.3
Pillow==10.2.0
python-docx==1.1.0
PyPDF2==3.0.1
cairosvg==2.7.1
markdown2==2.4.10
lxml==5.1.0